explicitly to long

+++ src/experiment_multiround/supervised_multi.py
@@ -54,7 +54,7 @@ class MultiroundExperiment(object):
         if round > 0:
             initial_pool = (self.dset['graph'].ndata['creation_round'] == 0).nonzero().flatten() if all_data else torch.Tensor([])
             positive_preds = torch.cat([torch.cat(self.rounds[i]['checks'][:2], 0) for i in (list(range(round)) if all_data else [round-1])], 0)
-            full_pool = torch.cat([initial_pool, positive_preds], 0)
+            full_pool = torch.cat([initial_pool, positive_preds], 0).long()
 
             if full_pool.shape[0] == 0:
                 return None
